+++ examples/examples/network/alice.rs
@@ -37,8 +37,6 @@ fn producer() -> Blueprint {
 }
 
 pub(crate) fn topology() -> elfo::Topology {
-    elfo::_priv::node::set_node_no(1);
-
     let topology = elfo::Topology::empty();
     let logger = elfo::batteries::logger::init();
     let telemeter = elfo::batteries::telemeter::init();